Fix bugs identified by GCC compiler warnings
- SckBase::createInfo(): add missing return true — function declared bool
  but fell off the end, causing undefined behavior
- Sck_SEN5X::idle(): add missing return true on the success path — same
  issue, function declared bool with no return on normal exit
- Sck_AS7331::writeByte(): return Wire.endTransmission() instead of
  discarding it — function declared byte but had no return statement
- SckLed::getColor() / getDiff(): add fallback return of first element if
  the requested color is not found, preventing undefined behavior on an
  unmatched lookup
- SckBase::sdPublish(): increase newName buffer from 16 to 20 bytes —
  the sprintf format can produce up to 19 chars (e.g. "25-12-31_255.CSV")
  which overflowed the 16-byte stack buffer
- SckBase::getRSSI(): fix %i format specifier to %lu for uint32_t argument
- Commands.cpp: fix %u format specifier to %lu for uint32_t readInterval

+++ sam/src/Commands.cpp
@@ -351,7 +351,7 @@ void sensorConfig_com(SckBase* base, String parameters)
             uint8_t newEveryNint = intervalInt / base->config.readInterval;
             if (newEveryNint < 1) newEveryNint = 1;
 
-            snprintf(base->outBuff, sizeof(base->outBuff), "The sensor read interval is calculated as a multiple of general read interval (%u)", base->config.readInterval);
+            snprintf(base->outBuff, sizeof(base->outBuff), "The sensor read interval is calculated as a multiple of general read interval (%lu)", (unsigned long)base->config.readInterval);
             base->sckOut();
             if (newEveryNint < 255) {
